+++ src/cli/address_gen.go
@@ -5,11 +5,11 @@ import (
 	"os"
 	"runtime"
 
-	messages "github.com/skycoin/hardware-wallet-protob/go"
+	messages "github.com/fibercrypto/skywallet-protob/go"
 
 	gcli "github.com/urfave/cli"
 
-	skyWallet "github.com/skycoin/hardware-wallet-go/src/skywallet"
+	skyWallet "github.com/fibercrypto/skywallet-go/src/skywallet"
 )
 
 func addressGenCmd() gcli.Command {
@@ -38,9 +38,15 @@ func addressGenCmd() gcli.Command {
 				Usage:  "Device type to send instructions to, hardware wallet (USB) or emulator.",
 				EnvVar: "DEVICE_TYPE",
 			},
+			gcli.StringFlag{
+				Name:     "walletType",
+				Usage:    "Wallet type. Types are \"deterministic\" or \"bip44\"",
+				Required: true,
+			},
 		},
 		OnUsageError: onCommandUsageError(name),
 		Action: func(c *gcli.Context) {
+			walletType := c.String("walletType")
 			addressN := c.Int("addressN")
 			startIndex := c.Int("startIndex")
 			confirmAddress := c.Bool("confirmAddress")
@@ -60,7 +66,7 @@ func addressGenCmd() gcli.Command {
 			}
 
 			var pinEnc string
-			msg, err := device.AddressGen(uint32(addressN), uint32(startIndex), confirmAddress)
+			msg, err := device.AddressGen(uint32(addressN), uint32(startIndex), confirmAddress, walletType)
 			if err != nil {
 				log.Error(err)
 				return