+++ src/cli/features.go
@@ -9,9 +9,9 @@ import (
 	"github.com/gogo/protobuf/proto"
 	gcli "github.com/urfave/cli"
 
-	messages "github.com/skycoin/hardware-wallet-protob/go"
+	messages "github.com/fibercrypto/skywallet-protob/go"
 
-	skyWallet "github.com/skycoin/hardware-wallet-go/src/skywallet"
+	skyWallet "github.com/fibercrypto/skywallet-go/src/skywallet"
 )
 
 func featuresCmd() gcli.Command {